Extract vectors by index or value occurrence

Extract front or rear of vector `x` up and untill an index `i`, the first or last occurence of a value `v`, or, extract values based on quantile `q`, first, middle, or, last index `j`.

Usage

x %[f% v

x %[l% v

x %[% i

x %]% i

x %f]% v

x %l]% v

x %q]% q

x %q)% q

x %[q% q

x %(q% q

x %:% j

Arguments

x

A vector

v

A value of which the first or last occurrence in x will be used as an index

i

An index or two element vector c(lo,hi) indicating a range to extract

q

A percentile value (between `0` and `1`)

j

A character indicating to extract the first `f`, middle `m` or last `l` value of `x`.

Value

A vector extracted from the front, rear, or, range of `x`. Either based on an index or the first or last occurrence of a value or the first, middle, or, ;ast element of a vector.

Note

The function provided for symmetry, character lengths of x%]%i and x[1:i] are equal.

Examples

z <- letters

# Extract front by first occurrence of value
z %[f% "n"
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n"

# Extract front by index
x <- rnorm(100)
x %[% 10
#>  [1] -1.05911107  0.21076504  0.01396083 -1.04393509  0.03218128 -0.14163579
#>  [7] -0.56953879 -0.33270695 -1.31592154 -1.82151090

# Extract rear by index
x %]% 90
#>  [1]  1.3825760  1.2238421  0.7718121 -0.4588196  1.2200944  0.7667757
#>  [7] -0.3190353 -1.2901086 -0.7827239 -0.3491098  0.1497596

# Extract rear by index
x %]% 90
#>  [1]  1.3825760  1.2238421  0.7718121 -0.4588196  1.2200944  0.7667757
#>  [7] -0.3190353 -1.2901086 -0.7827239 -0.3491098  0.1497596

# Extract by indices if a range is provided
x %]% c(4,30)
#>  [1] -1.04393509  0.03218128 -0.14163579 -0.56953879 -0.33270695 -1.31592154
#>  [7] -1.82151090  0.78659203 -2.50330474 -1.33534086  0.98244799  1.17610614
#> [13]  0.52626433 -0.09003325  1.08843791 -2.02696414  1.18429261  0.61854936
#> [19]  0.45963263  1.30944014 -0.13818654 -0.12052013  0.84485752 -0.19640971
#> [25] -0.03614495 -1.04873371  0.36412298
z %[% c(6,10)
#> [1] "f" "g" "h" "i" "j"

# Extract last/middle value of x
x %:% "l"
#> [1] 0.1497596
z %:% "m"
#> [1] "m"

# Extract by percentile
seq(1,10,.5) %(q% .5 # infix
#> [1] 1.0 1.5 2.0 2.5 3.0 3.5 4.0 4.5 5.0
seq(1,10,.5)[seq(1,10,.5) < quantile(seq(1,10,.5),.5)] # regular syntax
#> [1] 1.0 1.5 2.0 2.5 3.0 3.5 4.0 4.5 5.0

seq(1,10,.5) %q]% .5 # infix
#>  [1]  5.5  6.0  6.5  7.0  7.5  8.0  8.5  9.0  9.5 10.0
seq(1,10,.5)[seq(1,10,.5) >= quantile(seq(1,10,.5),.5)] # regular syntax
#>  [1]  5.5  6.0  6.5  7.0  7.5  8.0  8.5  9.0  9.5 10.0