Extract vectors by index or value occurrence

Extract front or rear of vector `x` up and untill an index `i`, the first or last occurence of a value `v`, or, extract values based on quantile `q`, first, middle, or, last index `j`.

x %[f% v

x %[l% v

x %[% i

x %]% i

x %f]% v

x %l]% v

x %q]% q

x %q)% q

x %[q% q

x %(q% q

x %:% j

Arguments

x
A vector

v
A value of which the first or last occurrence in x will be used as an index

i
An index or two element vector c(lo,hi) indicating a range to extract

q
A percentile value (between `0` and `1`)

j
A character indicating to extract the first `f`, middle `m` or last `l` value of `x`.

Value

A vector extracted from the front, rear, or, range of `x`. Either based on an index or the first or last occurrence of a value or the first, middle, or, ;ast element of a vector.

Note

The function provided for symmetry, character lengths of x%]%i and x[1:i] are equal.

Examples

z <- letters

# Extract front by first occurrence of value
z %[f% "n"

#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n"

# Extract front by index
x <- rnorm(100)
x %[% 10

#>  [1] -1.400043517  0.255317055 -2.437263611 -0.005571287  0.621552721
#>  [6]  1.148411606 -1.821817661 -0.247325302 -0.244199607 -0.282705449

# Extract rear by index
x %]% 90

#>  [1] -0.38195111  0.42418757  1.06310200  1.04871262 -0.03810289  0.48614892
#>  [7]  1.67288261 -0.35436116  0.94634789  1.31682636 -0.29664002

# Extract rear by index
x %]% 90

#>  [1] -0.38195111  0.42418757  1.06310200  1.04871262 -0.03810289  0.48614892
#>  [7]  1.67288261 -0.35436116  0.94634789  1.31682636 -0.29664002

# Extract by indices if a range is provided
x %]% c(4,30)

#>  [1] -0.005571287  0.621552721  1.148411606 -1.821817661 -0.247325302
#>  [6] -0.244199607 -0.282705449 -0.553699384  0.628982042  2.065024895
#> [11] -1.630989402  0.512426950 -1.863011492 -0.522012515 -0.052601910
#> [16]  0.542996343 -0.914074827  0.468154420  0.362951256 -1.304543545
#> [21]  0.737776321  1.888504929 -0.097445104 -0.935847354 -0.015950311
#> [26] -0.826788954 -1.512399651

z %[% c(6,10)

#> [1] "f" "g" "h" "i" "j"

# Extract last/middle value of x
x %:% "l"

#> [1] -0.29664

z %:% "m"

#> [1] "m"

# Extract by percentile
x %[q% .975

#>  [1] -1.400043517  0.255317055 -2.437263611 -0.005571287  0.621552721
#>  [6]  1.148411606 -1.821817661 -0.247325302 -0.244199607 -0.282705449
#> [11] -0.553699384  0.628982042 -1.630989402  0.512426950 -1.863011492
#> [16] -0.522012515 -0.052601910  0.542996343 -0.914074827  0.468154420
#> [21]  0.362951256 -1.304543545  0.737776321  1.888504929 -0.097445104
#> [26] -0.935847354 -0.015950311 -0.826788954 -1.512399651  0.935363190
#> [31]  0.176488611  0.243685465  1.623548883  0.112038083 -0.133997013
#> [36] -1.910087468 -0.279237242 -0.313445978  1.067307879  0.070034850
#> [41] -0.639123324 -0.049964899 -0.251483443  0.444797116  0.046531380
#> [46]  0.577709069  0.118194874 -1.911720491  0.862086482 -0.243236740
#> [51] -0.206087195  0.019177592  0.029560754  0.549827542 -2.274114857
#> [56] -0.361221255  0.213355750  1.074345882 -0.665088249  1.113952419
#> [61] -0.245896412 -1.177563309 -0.975850616  1.065057320  0.131670635
#> [66]  0.488628809 -1.699450568 -1.470736306  0.284150344  1.337320413
#> [71]  0.236696283  1.318293384  0.523909788  0.606748047 -0.109935672
#> [76]  0.172181715 -0.090327287  1.924343341  1.298392759  0.748791268
#> [81]  0.556224329 -0.548257264  1.110534893 -2.612334333 -0.155693776
#> [86]  0.433889790 -0.381951112  0.424187575  1.063101996  1.048712620
#> [91] -0.038102895  0.486148920  1.672882611 -0.354361164  0.946347886
#> [96]  1.316826356 -0.296640025

x %q)% .025

#>  [1] -1.400043517  0.255317055 -0.005571287  0.621552721  1.148411606
#>  [6] -1.821817661 -0.247325302 -0.244199607 -0.282705449 -0.553699384
#> [11]  0.628982042  2.065024895 -1.630989402  0.512426950 -1.863011492
#> [16] -0.522012515 -0.052601910  0.542996343 -0.914074827  0.468154420
#> [21]  0.362951256 -1.304543545  0.737776321  1.888504929 -0.097445104
#> [26] -0.935847354 -0.015950311 -0.826788954 -1.512399651  0.935363190
#> [31]  0.176488611  0.243685465  1.623548883  0.112038083 -0.133997013
#> [36] -1.910087468 -0.279237242 -0.313445978  1.067307879  0.070034850
#> [41] -0.639123324 -0.049964899 -0.251483443  0.444797116  2.755417575
#> [46]  0.046531380  0.577709069  0.118194874 -1.911720491  0.862086482
#> [51] -0.243236740 -0.206087195  0.019177592  0.029560754  0.549827542
#> [56]  2.682557184 -0.361221255  0.213355750  1.074345882 -0.665088249
#> [61]  1.113952419 -0.245896412 -1.177563309 -0.975850616  1.065057320
#> [66]  0.131670635  0.488628809 -1.699450568 -1.470736306  0.284150344
#> [71]  1.337320413  0.236696283  1.318293384  0.523909788  0.606748047
#> [76] -0.109935672  0.172181715 -0.090327287  1.924343341  1.298392759
#> [81]  0.748791268  0.556224329 -0.548257264  1.110534893 -0.155693776
#> [86]  0.433889790 -0.381951112  0.424187575  1.063101996  1.048712620
#> [91] -0.038102895  0.486148920  1.672882611 -0.354361164  0.946347886
#> [96]  1.316826356 -0.296640025