Regress vectors

Regress vectors

Correlate x and y

Polynomial regression of degree 1

Polynomial regression of degree 2

Polynomial regression of degree 3

Polynomial regression of degree 4

`

Usage

x %/r% y

x %/1% y

x %/2% y

x %/3% y

x %/4% y

x %/n% yn

Arguments

x

Numeric vectors

y

Numeric vector

yn

List of length 2, first element is a vector y, the second element an integer denoting the order of the polynomial regression.

Examples

x <- rnorm(100)
y <- x + x^2 + x^3

# Correlate x with y
x%/r%y
#> [1] 0.8884375

# Polynomial regression degree 1 .. 4
x%/1%y
#> 
#> Call:
#> stats::lm(formula = y ~ stats::poly(x, order = 1))
#> 
#> Coefficients:
#>               (Intercept)  stats::poly(x, order = 1)  
#>                    0.4576                    25.3894  
#> 
x%/2%y
#> 
#> Call:
#> stats::lm(formula = y ~ stats::poly(x, order = 2))
#> 
#> Coefficients:
#>                (Intercept)  stats::poly(x, order = 2)1  
#>                     0.4576                     25.3894  
#> stats::poly(x, order = 2)2  
#>                     2.5157  
#> 
x%/3%y
#> 
#> Call:
#> stats::lm(formula = y ~ stats::poly(x, order = 3))
#> 
#> Coefficients:
#>                (Intercept)  stats::poly(x, order = 3)1  
#>                     0.4576                     25.3894  
#> stats::poly(x, order = 3)2  stats::poly(x, order = 3)3  
#>                     2.5157                     12.8736  
#> 
x%/4%y
#> 
#> Call:
#> stats::lm(formula = y ~ stats::poly(x, order = 4))
#> 
#> Coefficients:
#>                (Intercept)  stats::poly(x, order = 4)1  
#>                  4.576e-01                   2.539e+01  
#> stats::poly(x, order = 4)2  stats::poly(x, order = 4)3  
#>                  2.516e+00                   1.287e+01  
#> stats::poly(x, order = 4)4  
#>                 -2.068e-15  
#> 

anova(x%/1%y,x%/2%y,x%/3%y,x%/4%y)
#> Analysis of Variance Table
#> 
#> Model 1: y ~ stats::poly(x, order = 1)
#> Model 2: y ~ stats::poly(x, order = 2)
#> Model 3: y ~ stats::poly(x, order = 3)
#> Model 4: y ~ stats::poly(x, order = 4)
#>   Res.Df    RSS Df Sum of Sq          F  Pr(>F)    
#> 1     98 172.06                                    
#> 2     97 165.73  1     6.329 4.3861e+30 < 2e-16 ***
#> 3     96   0.00  1   165.729 1.1486e+32 < 2e-16 ***
#> 4     95   0.00  1     0.000 2.9634e+00 0.08842 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

# Order n

x%/n%list(y,10)
#> 
#> Call:
#> stats::lm(formula = yn[[1]] ~ stats::poly(x, order = yn[[2]]))
#> 
#> Coefficients:
#>                       (Intercept)   stats::poly(x, order = yn[[2]])1  
#>                         4.576e-01                          2.539e+01  
#>  stats::poly(x, order = yn[[2]])2   stats::poly(x, order = yn[[2]])3  
#>                         2.516e+00                          1.287e+01  
#>  stats::poly(x, order = yn[[2]])4   stats::poly(x, order = yn[[2]])5  
#>                        -2.068e-15                          8.807e-16  
#>  stats::poly(x, order = yn[[2]])6   stats::poly(x, order = yn[[2]])7  
#>                        -9.857e-16                         -4.085e-16  
#>  stats::poly(x, order = yn[[2]])8   stats::poly(x, order = yn[[2]])9  
#>                        -3.216e-16                         -2.718e-16  
#> stats::poly(x, order = yn[[2]])10  
#>                         1.442e-15  
#>